Breaking out a replica applier type from the replica controller type
