+++ go/cmd/dolt/commands/engine/sqlengine.go
@@ -142,7 +142,8 @@ func NewSqlEngine(
 		"authentication_dolt_jwt": NewAuthenticateDoltJWTPlugin(config.JwksConfig),
 	})
 
-	engine.BinlogReplicaController = config.BinlogReplicaController
+	// TODO: This could be a cleaner interface
+	engine.Analyzer.BinlogReplicaController = config.BinlogReplicaController
 
 	// Load MySQL Db information
 	if err = engine.Analyzer.Catalog.MySQLDb.LoadData(sql.NewEmptyContext(), data); err != nil {
